+++ samples/sqlservices/src/featureProviders/iconProvider.ts
